+++ src/Algorithm.hpp
@@ -3,12 +3,12 @@
 class Algorithm {
     private:
         int size;
-        int **array;
+        unsigned short **array;
     public:
         explicit Algorithm(int size);
         ~Algorithm();
-        void Set(int index1, int index2, int value);
-        int Get(int index1, int index2);
+        void Set(int index1, int index2, unsigned short value);
+        unsigned short Get(int index1, int index2);
         //void PrintLength();
  
 };
